* Updates and fixes to house pages and controllers * Minor cosmetic changes on cursed object pages

+++ front-end-and-rest-layer/src/main/webapp/haunted_houses_app.js
@@ -437,15 +437,10 @@ hauntedHousesControllers.controller('AbilityUpdateController', function ($scope,
 
 /*** Cursed object controllers ***/
 
-// Loads cursed objects
-function loadAllCursedObjects($http, $scope) {
+hauntedHousesControllers.controller('CursedObjectController', function ($scope, $rootScope, $location, $routeParams, $http) {
     $http.get('rest/cursedObjects').then(function (response) {
         $scope.cursedObjects = response.data;
     });
-}
-
-hauntedHousesControllers.controller('CursedObjectController', function ($scope, $rootScope, $location, $routeParams, $http) {
-    loadAllCursedObjects($http, $scope);
     
     $http.get('rest/houses').then(function (response) {
         $scope.houses = response.data;
@@ -467,7 +462,7 @@ hauntedHousesControllers.controller('CursedObjectController', function ($scope, 
             url: 'rest/cursedObjects/increase',
             data: threshold
         }).then(function success() {
-            $rootScope.successAlert = 'Monster attraction factor was successfully increased on specified cursed objects.';
+            $rootScope.successAlert = 'Monster attraction factor was successfully increased on cursed objects below threshold.';
             $location.path("/cursedObjects");
         }, function error(response) {
             console.log("Error when increasing monster attraction factor.");
@@ -587,28 +582,51 @@ hauntedHousesControllers.controller('CursedObjectUpdateController', function ($s
 /*** House controllers ***/
 
 hauntedHousesControllers.controller('HousesController', function ($scope, $http, $rootScope, $location) {
-    $http.get('/pa165/rest/houses').then(function (response) {
+    $http.get('rest/houses').then(function (response) {
         $scope.houses = response.data;
     });
+    
+    $http.get('rest/monsters').then(function (response) {
+        $scope.monsters = response.data;
+    });
+    
+    $http.get('rest/cursedObjects').then(function (response) {
+        $scope.cursedObjects = response.data;
+    });
 
+    $scope.filterByMonsterId = function (house) {
+        return function (monster) {
+            return house.monsterIds.indexOf(monster.id) !== -1;
+        };
+    };
+    
+    $scope.filterByCursedObjectId = function (house) {
+        return function (cursedObject) {
+            return house.cursedObjectIds.indexOf(cursedObject.id) !== -1;
+        };
+    };
+    
     $scope.delete = function(house) {
-        console.log("Deleting house with id = " + house.id + " (" + house.name + ")");
+        if (!confirm("Are you certain you want to delete chosen house?")) {
+            return;
+        }
+        
+        console.log("Deleting house with id: " + house.id);
         $http.delete("rest/houses/" + house.id).then(
                 function success(response) {
-                    console.log("Succesfully deleted house " + house.id + " on the server");
-                    // Display confirmation alert
-                    $rootScope.successAlert = ("Deleted house \"" + house.name + "\"");
+                    console.log("Succesfully deleted house " + house.id + " on the server.");
+                    $rootScope.successAlert = 'Deleted house: "' + house.name + '"';
                     $location.path("/houses");
                 },
                 function error(response) {
-                    console.log("Error when deleting house with id \"" + house.id + "\"");
+                    console.log("Error when deleting house with id: " + house.id);
                     console.log(response);
                     switch (response.data.code) {
                         case 'ResourceNotFoundException':
-                            $rootScope.errorAlert = "Cannot delete non-existent ability!";
+                            $rootScope.errorAlert = "Cannot delete non-existent house!";
                             break;
                         default:
-                            $rootScope.errorAlert = "Cannot delete house! Reason given by the server: " + response.data.message;
+                            $rootScope.errorAlert = "Cannot delete house with assigned monsters or cursed objects!";
                             break;
                     }
                 }
@@ -618,11 +636,11 @@ hauntedHousesControllers.controller('HousesController', function ($scope, $http,
 
 hauntedHousesControllers.controller('HouseCreateController', function ($scope, $http, $rootScope, $location) {
 
-    $http.get('/pa165/rest/monsters').then(function (response) {
+    $http.get('rest/monsters').then(function (response) {
         $scope.monsters = response.data;
     });
 
-    $http.get('/pa165/rest/cursedObjects').then(function (response) {
+    $http.get('rest/cursedObjects').then(function (response) {
         $scope.cursedObjects = response.data;
     });
 
@@ -633,24 +651,21 @@ hauntedHousesControllers.controller('HouseCreateController', function ($scope, $
         'cursedObjectIds': []
     };
 
-    // Create button clicked
     $scope.create = function (house) {
         house.monsterIds = getIdsFromSelection($scope.monsters);
         house.cursedObjectIds = getIdsFromSelection($scope.cursedObjects);
 
         $http({
             method: 'POST',
-            url: '/pa165/rest/houses/create',
+            url: 'rest/houses/create',
             data: house
         }).then(function success(response) {
-
             var createdHouse = response.data;
-            $rootScope.successAlert = "house \"" + createdHouse.name + "\" was created.";
+            $rootScope.successAlert = 'House with name "' + createdHouse.name + '" was created.';
             console.log("house " + createdHouse.name + " created");
             $location.path("/houses");
 
         }, function error(response) {
-
             console.log("Error when attempting to create house:");
             console.log(house);
             console.log(response);
@@ -667,7 +682,7 @@ hauntedHousesControllers.controller('HouseCreateController', function ($scope, $
 });
 
 hauntedHousesControllers.controller('HouseUpdateController', function ($scope, $http, $routeParams, $rootScope, $location) {
-    $http.get('/pa165/rest/monsters').then(function (response) {
+    $http.get('rest/monsters').then(function (response) {
         $scope.monsters = response.data;
 
         // Clear checked status
@@ -685,7 +700,7 @@ hauntedHousesControllers.controller('HouseUpdateController', function ($scope, $
         }
     });
 
-    $http.get('/pa165/rest/cursedObjects').then(function (response) {
+    $http.get('rest/cursedObjects').then(function (response) {
         $scope.cursedObjects = response.data;
 
         // Clear checked status
@@ -710,9 +725,9 @@ hauntedHousesControllers.controller('HouseUpdateController', function ($scope, $
         'cursedObjectIds': []
     };
 
-    //get object that should be updated
+    // Get object that should be updated
     var houseId = $routeParams.houseId;
-    $http.get('/pa165/rest/houses/id/' + houseId).then(function (response) {
+    $http.get('rest/houses/id/' + houseId).then(function (response) {
         var house = response.data;
         $scope.house = house;
     });
@@ -724,12 +739,11 @@ hauntedHousesControllers.controller('HouseUpdateController', function ($scope, $
 
         $http({
             method: 'PUT',
-            url: '/pa165/rest/houses/' + $scope.house.id,
+            url: 'rest/houses/' + $scope.house.id,
             data: house
         }).then(function success(response) {
-
             var updatedHouse = response.data;
-            $rootScope.successAlert = "house \"" + updatedHouse.name + "\" was updated.";
+            $rootScope.successAlert = 'House with name "' + updatedHouse.name + '" was updated.';
             console.log("house " + updatedHouse.name + " updated");
             $location.path("/houses");
